+++ lib/src/parser/models/flavors/linux.g.dart
@@ -1,0 +1,16 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'linux.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+Linux _$LinuxFromJson(Map json) {
+  $checkKeys(
+    json,
+    requiredKeys: const ['applicationId'],
+    disallowNullValues: const ['applicationId'],
+  );
+  return Linux(applicationId: json['applicationId'] as String);
+}